fixed a bug with posting after making the adjustement for bug #106
